move createConfigfile into separate file
